requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.12
Werkzeug==3.0.1
pandas>=2.2.2
numpy>=2.0.0

